requirements.txt
Django>=4.0,<5.0
djangorestframework>=3.12,<4.0
Pillow>=9.0,<10.0
